complete homework_3
