+++ NailBot/Program.cs
@@ -51,23 +51,17 @@ namespace NailBot
                 catch (Exception ex)
                 {
                     Console.WriteLine($"Основное исключение: {ex.Message}\n");
-                    Console.WriteLine($"Стек трейс основного исключения: {ex.StackTrace}\n");
+                    //Console.WriteLine($"Стек трейс основного исключения: {ex.StackTrace}\n");
                     if (ex.InnerException != null)
                     {
                         Console.WriteLine($"Внутреннее исключение: {ex.InnerException.Message}\n");
-                        Console.WriteLine($"Источник внутреннего исключения: {ex.InnerException.StackTrace}\n");
+                        //Console.WriteLine($"Источник внутреннего исключения: {ex.InnerException.StackTrace}\n");
                     }
                 }
             }
 
             //DescriptionBot.PrintDescription();
 
-            // не совсем до конца понял по количеству try catch и как будто бы переборщил с ними))) 
-            // повешал глобальный try catch внутри которого ещё try catch которые отлавливают исключения из задания
-            // и выбрасывают ещё одно, чтобы прочитать InnerException, всё время получал null, потому так обвешал
-
-            // не совсем понял каким образом отловить Exception и вывести InnerException потому что во всех случаях ввода предусматриваются те или иные кастомные Exception
-            // по заданию там нужно выводить только сообщение об ошибке, без innerException
         }
     }
 }
